requirements.txt
python-telegram-bot==20.3
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0